better handling for already existing output folder and/or posts folders

+++ main.ts
@@ -1,6 +1,5 @@
 import { readXml } from './src/readXml.ts';
 import { processPost } from './src/processPost.ts';
-import { parse } from '@libs/xml';
 
 async function main() {
     const [inputFilePath = 'export.xml'] = Deno.args;
@@ -9,12 +8,21 @@ async function main() {
     const { link, posts, attachments } = await readXml(inputFilePath);
     console.log('Wordpress XML has been read and parsed!');
 
-    // todo prompt if the user really wants to delete the output folder
-
-    console.log('Deleting & recreating ./output...');
-    await Deno.remove('output', { recursive: true });
-    await Deno.mkdir('output');
-    console.log('Fresh output folder is ready!');
+    const fileInfo = await Deno.stat('output').catch(() => {});
+    if (fileInfo) {
+        if (fileInfo.isFile) {
+            console.error('%cA file named "output" already exists; exiting.', 'color: red');
+            Deno.exit(1);
+        }
+        if (fileInfo.isDirectory) {
+            const deleteOutput = confirm('The folder "output" already exists; do you want to delete it first');
+            if (deleteOutput) {
+                await Deno.remove('output', { recursive: true });
+                await Deno.mkdir('output');
+                console.log('Fresh output folder is ready!');
+            }
+        }
+    }
 
     for (const post of posts) {
         await processPost(post, link, attachments);
